Remove cluster_metadata.go
