+++ temporal/server.go
@@ -30,9 +30,11 @@ import (
 	"sync"
 	"time"
 
+	"github.com/pborman/uuid"
 	"github.com/uber-go/tally"
 	sdkclient "go.temporal.io/sdk/client"
 
+	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/archiver"
 	"go.temporal.io/server/common/archiver/provider"
@@ -43,6 +45,7 @@ import (
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
+	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/cassandra"
 	persistenceClient "go.temporal.io/server/common/persistence/client"
 	"go.temporal.io/server/common/persistence/elasticsearch/client"
@@ -58,6 +61,10 @@ import (
 	"go.temporal.io/server/service/history"
 	"go.temporal.io/server/service/matching"
 	"go.temporal.io/server/service/worker"
+)
+
+const (
+	mismatchLogMessage = "Supplied configuration key/value mismatches persisted cluster metadata. Continuing with the persisted value as this value cannot be changed once initialized."
 )
 
 type (
@@ -357,6 +364,76 @@ func verifyPersistenceCompatibleVersion(config config.Persistence, persistenceSe
 		return fmt.Errorf("sql schema version compatibility check failed: %w", err)
 	}
 	return nil
+}
+
+// initClusterMetadata performs a config check against the configured persistence store for cluster metadata.
+// If there is a mismatch, the persisted values take precedence and will be written over in the config objects.
+// This is to keep this check hidden from downstream call although they should not use config directly.
+func initClusterMetadata(cfg *config.Config, persistenceServiceResolver resolver.ServiceResolver, logger log.Logger) (cluster.Metadata, error) {
+	logger = log.With(logger, tag.ComponentMetadataInitializer)
+
+	factory := persistenceClient.NewFactory(
+		&cfg.Persistence,
+		persistenceServiceResolver,
+		nil,
+		nil,
+		cfg.ClusterMetadata.CurrentClusterName,
+		nil,
+		logger,
+	)
+
+	clusterMetadataManager, err := factory.NewClusterMetadataManager()
+	if err != nil {
+		return nil, fmt.Errorf("error initializing cluster metadata manager: %w", err)
+	}
+	defer clusterMetadataManager.Close()
+
+	applied, err := clusterMetadataManager.SaveClusterMetadata(
+		&persistence.SaveClusterMetadataRequest{
+			ClusterMetadata: persistencespb.ClusterMetadata{
+				HistoryShardCount: cfg.Persistence.NumHistoryShards,
+				ClusterName:       cfg.ClusterMetadata.CurrentClusterName,
+				ClusterId:         uuid.New(),
+			}})
+	if err != nil {
+		logger.Warn("Failed to save cluster metadata.", tag.Error(err))
+	}
+	if applied {
+		logger.Info("Successfully saved cluster metadata.")
+	} else {
+		resp, err := clusterMetadataManager.GetClusterMetadata()
+		if err != nil {
+			return nil, fmt.Errorf("error while fetching cluster metadata: %w", err)
+		}
+		if cfg.ClusterMetadata.CurrentClusterName != resp.ClusterName {
+			logger.Error(
+				mismatchLogMessage,
+				tag.Key("clusterMetadata.currentClusterName"),
+				tag.IgnoredValue(cfg.ClusterMetadata.CurrentClusterName),
+				tag.Value(resp.ClusterName))
+			cfg.ClusterMetadata.CurrentClusterName = resp.ClusterName
+		}
+
+		var persistedShardCount = resp.HistoryShardCount
+		if cfg.Persistence.NumHistoryShards != persistedShardCount {
+			logger.Error(
+				mismatchLogMessage,
+				tag.Key("persistence.numHistoryShards"),
+				tag.IgnoredValue(cfg.Persistence.NumHistoryShards),
+				tag.Value(persistedShardCount))
+			cfg.Persistence.NumHistoryShards = persistedShardCount
+		}
+	}
+
+	metadata := cluster.NewMetadata(
+		logger,
+		cfg.ClusterMetadata.EnableGlobalNamespace,
+		cfg.ClusterMetadata.FailoverVersionIncrement,
+		cfg.ClusterMetadata.MasterClusterName,
+		cfg.ClusterMetadata.CurrentClusterName,
+		cfg.ClusterMetadata.ClusterInformation,
+	)
+	return metadata, nil
 }
 
 func initSystemNamespaces(cfg *config.Persistence, currentClusterName string, persistenceServiceResolver resolver.ServiceResolver, logger log.Logger) error {
